rename tantalus file to bnb and fix IMU data length
